requirements.txt
Flask
Flask-SQLAlchemy
pymongo
google-cloud-firestore
paho-mqtt
python-dotenv
firebase-admin
bson
plotly

Flask-SQLAlchemy
Flask
Flask-SQLAlchemy
pymongo
google-cloud-firestore
paho-mqtt
python-dotenv
firebase-admin
bson
plotly

pandas
openpyxl
flask-wtf
flask-bootstrap
